Added ESpeed component! 🥳
